refactor: user vm factory
