+++ ReimbifyApp/app/src/main/java/com/example/reimbifyapp/user/factory/UserViewModelFactory.kt
@@ -1,13 +1,13 @@
-package com.example.reimbifyapp.user.viewmodel
+package com.example.reimbifyapp.user.factory
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.example.reimbifyapp.user.data.repositories.UserRepository
 import com.example.reimbifyapp.user.di.Injection
-import com.example.reimbifyapp.user.ui.AuthActivity
+import com.example.reimbifyapp.user.viewmodel.LoginViewModel
 
-class ViewModelFactory(
+class UserViewModelFactory(
     private val userRepository: UserRepository,
 ) : ViewModelProvider.NewInstanceFactory() {
 
@@ -23,13 +23,13 @@ class ViewModelFactory(
 
     companion object {
         @Volatile
-        private var INSTANCE: ViewModelFactory? = null
+        private var INSTANCE: UserViewModelFactory? = null
 
         @JvmStatic
-        fun getInstance(context: Context): ViewModelFactory {
+        fun getInstance(context: Context): UserViewModelFactory {
             return INSTANCE ?: synchronized(this) {
                 val userRepository = Injection.provideUserRepository(context)
-                INSTANCE = ViewModelFactory(userRepository)
+                INSTANCE = UserViewModelFactory(userRepository)
                 INSTANCE!!
             }
         }
