+++ ReimbifyApp/app/src/main/java/com/example/reimbifyapp/user/ui/ForgotPasswordFragment.kt
@@ -2,34 +2,26 @@ package com.example.reimbifyapp.user.ui
 
 import android.animation.AnimatorSet
 import android.animation.ObjectAnimator
-import android.content.Intent
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
 import android.widget.Toast
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
-import com.example.reimbifyapp.MainActivityUser
 import com.example.reimbifyapp.R
 import com.example.reimbifyapp.databinding.FragmentForgotPasswordBinding
-import com.example.reimbifyapp.databinding.FragmentLoginBinding
-import com.example.reimbifyapp.user.data.entities.User
 import com.example.reimbifyapp.user.utils.ErrorUtils.parseErrorMessage
 import com.example.reimbifyapp.user.viewmodel.LoginViewModel
-import com.example.reimbifyapp.user.viewmodel.ViewModelFactory
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.launch
+import com.example.reimbifyapp.user.factory.UserViewModelFactory
 
 class ForgotPasswordFragment : Fragment(R.layout.fragment_forgot_password) {
     private var _binding: FragmentForgotPasswordBinding? = null
     private val binding get() = _binding!!
 
     private val viewModel by viewModels<LoginViewModel> {
-        ViewModelFactory.getInstance(requireActivity())
+        UserViewModelFactory.getInstance(requireActivity())
     }
 
     override fun onCreateView(
